refactor(cli): isolate the uipath-ipc contract into its own module [ROBO-5779]
Move the wire-dictated PascalCase surface (PythonRunRequest/PythonRunResult,
IPythonRuntimeServer, PythonRuntimeService, start_ipc_server) out of cli_server
into cli_server_ipc, re-exported for backward compatibility. Scopes the Sonar
S100/S116 suppression to that single file so the rest of the tree stays strict.
