+++ sonar-project.properties
@@ -12,4 +12,16 @@ sonar.exclusions=**/samples/**,**/testcases/**,**/template/**,**/_resources/**
 
 sonar.cpd.exclusions=**/__init__.py
 
+# The uipath-ipc runtime contract (methods + DTO fields) is PascalCase because
+# the wire keys and method names are dictated by the .NET/CoreIpc peer — the
+# serializer maps them verbatim (no alias mechanism), so the names cannot be
+# Python-idiomatic without breaking interop. It is isolated in cli_server_ipc.py
+# so this suppression of method-naming (S100) and field-naming (S116) touches
+# that file only.
+sonar.issue.ignore.multicriteria=ipc1,ipc2
+sonar.issue.ignore.multicriteria.ipc1.ruleKey=python:S100
+sonar.issue.ignore.multicriteria.ipc1.resourcePath=**/cli_server_ipc.py
+sonar.issue.ignore.multicriteria.ipc2.ruleKey=python:S116
+sonar.issue.ignore.multicriteria.ipc2.resourcePath=**/cli_server_ipc.py
+
 sonar.sourceEncoding=UTF-8
